backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0
razorpay==1.3.0
Pillow==10.4.0
reportlab==4.0.4
cloudinary==1.44.1
gunicorn==21.2.0 